adjustments from dimok's latest HBL rpx to try to get sound working
Compiles to an RPX, but crashes instantly. Needs dimok's branch of
WUT. Next step is probably to debug it using the UDP logger to see
where it crashes. I based this branch on the hbl_elf branch, and also
tried to merge in some changes dimok made to HBL and some changes from
the normal, soundless C wut branch.